IB/hfi: Capability bitmask and user level header
This commit contains two different changes:

1. Capability bitmask module parameter

As part of the effort to switch control of HW and
driver feature to the driver (and remove the matching
code from PSM), a new set of feature bits is being
introduced.

These bit are paired with a new module parameter, which
is used to enable/disable HW and driver features. The
new module parameter is a bitmask with the following
format:

      | |62         |47         |23           |0
      +-+-----------+-----------+-------------+
      |L| Misc Bits | User Bits | Kernel Bits |
      +-+-----------+-----------+-------------+

 * "L" is a single bit that indicates the bitmask has
   been "locked". Locking of the bitmask happens when
   the driver probe function is called and prevents the
   user from changing capabilities that require setup
   during driver and HW initialization.
 * "Misc Bits" are 15 bits that can be used to communicate
   specific behavior or workarounds (not corresponding to
   HW or driver features) to user level processes.
 * "User Bits" are the bits used as the capability bitmask
   for user contexts.
 * "Kernel Bits" are the bits used as the capability
   bitmask for kernel contexts.

For the most part the "Kernel bits" and "User bits" can
have their own combination of feature bits. There are a
few exceptions for features that only make sense on a
"global" driver scale (not per-context or per-context type).
However, these "global" features have to be communicated to
user level processes. These features are contained in the
HFI_CAP_K2U mask (for kernel-to-user).

To ease the transition between the old and new module
parameters, the old boolean parameters have been
converted to "dummy" parameters which don't hold values
but rather update the new capability mask parameter
directly.

2. User level header

The driver defines an API for user level libraries used to
allocate contexts and resources. Previously, both the driver
and PSM have had their own header file which contained the
same set of definitions.

This change introduces a new header file (hfi_user.h), which
contains all the user level definitions (including the changes
described above).

This new header file will be "exported" to user level and user
level libraries should be including it instead of maintaining
their own copy.

INTERNAL: PR 124592 - PSM/driver API change for drop #6

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I12ac0b4be5e732618d432baafee6d346996824f1