IB/hfi: initialize 'total_cct_entry' element of struct qib_pportdata
The initialization of the 'total_cct_entry' element of struct
qib_pportdata was removed in commit 8fbde892. 'total_cct_entry' is
used in the read_cc_table_bin() method, and if it is not properly
initialized that method won't return the entire table.

INTERNAL: PR 125838 - initialize 'total_cct_entry' element of struct
INTERNAL: PR 125838 - qib_pportdata

Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>