IB/hfi: Workaround for Erratum 291241 - Credit reallocation
Add the workaround required for erratum 291241 in the dynamic
credit reallocation algorithm.

The workaround is to always reduce the shared credits to zero when
adjusting if any shared credit value is changing.

It is unknown if the erratum will be fixed for B0, so leave
a TODO to be resolved when the final decision is known.

INTERNAL: 125618 - Erratum 291241: Modifications to credit reallocation
INTERNAL: algorithm when per-VL shared limits are changing

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>