IB/hfi: Don't clobber important bits on AHG update
The driver was mistakenly clobbering the KDETH.SH and
KDETH.Intr bits with the AHG update of the KDETH.TIDCtrl
and KDETH.TID fields because of a buggy bit width value
in the AHG descriptor.

This was causing the HW to deliver headers for middle
packets (header, which should have been suppressed). In
turn, PSM was seeing packets that it wasn't expecting
which cause it to go into error state and re-transmit.

The end result of this was bad bandwidth and "hangs" in
PSM.

This commit fixes the number of bits updated in the AHG
descriptor so bits that don't need to be updated are left
alone.

INTERNAL: PR 125508 - Automatic Header Generation (AHG) for PSM

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>