Ib/hfi: Snoop regression test needs to use correct dest qpn
The snoop regression test assumes it can just flip the dlid/slid around in a
packet to send it back to the source, but this is not correct. It should be able
to set the correct dest qp.

This change also adds a subset of the snoop tests to the quick tests. Snoop is
important enough of a feature we need to ensure it is always working, the more
coverage the better. The additional tests do not add much time either. The more
intrusive tests like the IOCTL test are still exempt from quick testing.

INTERNAL: PR 125851 - Snoop.py regression test fails with ifs_fm version 186

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: I6b79dfca2a3be17c07ef05c7c641420e6438253d