IB/hfi: initial CCA implementation for verbs
This patch adds support for "Congestion Control Architecture" (CCA)
as defined in the IB volume 1 spec, release 1.2.1, Annex A10. It uses
WFR's "static rate control" mechanism (see WFR HAS section 7.4.5) to
introduce intra-packet delays on a link when any SL on that link is
congested. (A mechanism which provides per-SL rate control would be
preferable, but there's no hardware support for that in WFR.)
Generally, the receipt of a BECN causes the intra-packet delay to
rise, and periodically the delay is reduced. High resolution timers
are used to schedule delay reductions.

With RC QPs, BECNs are returned in ACK packets. For UD, and UC,
Congestion Notification Packets (CNPs) carry BECNs. Note that CNPs
use an opcode which is unrecognized by WFR A0 hardware (see WFR HAS,
section 8.4.2). The consequence of this is that CNPs are treated as
"error packets", and delivered to context 0 for processing.

INTERNAL: PR 125617 - CCA support for verbs

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: Ia663957fd56f5fdbb96004b8837b55604d2eb99a