nd: set module param defaults to match Simics+LBG
The platform that we currently care about is Simics with the LBG BIOS,
which is the platform that Validation is using to test all of our code.
This is also the platform that Manageability uses for integration, etc.
Unfortunately this platform requires a bunch of workarounds for out of
date NFIT data, etc.  Currently our workarounds are in the code, but
require module parameters passed at the command line or at the grub
kernel command line.

Since the only purpose of these workarounds is to support our
development platform and pretty much everyone running our code are using
this platform, just turn them on by default so we don't have to pass in
kernel module params.  All of this (including the defaults I'm setting)
will be removed when our development platforms support a current NFIT.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>