nd_dsm: return EINVAL on IOCTL errors
The old code was returning the positive status found in the mailbox,
which wasn't meaningful to Linux and was not being interpreted as an
error.

Also, even in the face of a mailbox error we want to copy the output
payload into the user buffer so that the user can check out the vendor
specific return codes for more information on the failure.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>