nd_dsm: All large payload support
This adds large payload support for manual DSM commands, as described in
the "NVDIMM FW Interface Table, Intel Specific Command Interface" spec
from Rob Branch.  This code is consistent with the 0.50 version of the
spec from October 20, 2014.

The only think of note is that the length of the input and output
buffers for vendor specific BIOS commands is specified twice, once as
part of the larger parent structure defined in section 2, and once as
part of the internal sub-structures defined in section 5.3.  I've mailed
Rob Branch about this, but for simplicity in this version of the code I
just ignore the inner size and use the size specified in the parent
structure.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>