nd_acpi: get rid of WARN_TAINT_ONCE for old NFIT
Right now every time a user loads nd_acpi and it detects an old NFIT, it
dumps a stack trace in syslog.  In my opinion this just distracts the
user, and serves no real purpose.  The driver can't control the NFIT
that is given to it, so just work gracefully in all cases.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>