IB/hfi: Don't move to Polling  when 8051 setup commands fail
Some wrappers of do_8051_command did not propagate error
values up the call tree, resulting in the driver moving to
polling state even when the 8051 setup commands failed. This
patch prevents continuing with LNI when the 8051 hardware is
non-responsive to host commands.

INTERNAL: PR 125337 - driver thinks it is polling even when
INTERNAL: the 8051 set-up commands fail

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Easwar Hariharan <easwar.hariharan@intel.com>
Change-Id: Ia83b4448952f1474c2e01579ab458e20260ac2d7