nd_dsm: remove redundant defines
Previously we had multiple defines for 1MB (CR_IN_MB_SIZE and
CR_OUT_MB_SIZE) and for the payload register sizes (CR_IN_PAYLOAD_SIZE
and CR_OUT_PAYLOAD_SIZE).  These are always the same, so get rid of the
IN/OUT definitions and just use a common define.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>