nd_dsm: correctly report _DSM status
Previously we returned bad IOCTL status (-EINVAL) when the IOCTL was
sent to the AEP firmware successfully, but that the AEP firmware wanted
to return some bad status like invalid opcode, etc.

Instead, the right thing to do is to return good IOCTL status in these
cases, and let the user pull the bad AEP status, and possibly the
extended status, out of the return buffer.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>