nd: minor cleanups after folding the dimm_radix change into the upstream patch set
Using the async domain for dimm deletion means we introduce portions of
the async infrastructure earlier in the patch set.  In turn this caused
me to recognize that include/linux/nd.h was needlessly including
async.h.  It also seemed odd to keep calling the domain
nd_async_register now that it has grown a usage outside of registration.