nd: add interleave-set state-tracking infrastructure
On nvdimms that support internal labeling, a portion of the dimm may be
accessible via an interleave set PMEM mapping in addition to the dimm's
BLK (block-data-window aperture(s)) interface.  A label, stored in a
"configuration data region" on the dimm, disambiguates which dimm
addresses are accessed through which exclusive interface.

Add infrastructure that allows the kernel to block modifications to a
label in the set while any member is active.  Note that this is meant
only for enforcing "no modifications of active labels" via the coarse
ioctl command.  The label modifications needed to support adding and
deleting namespaces will not require an inactive interleave set.
However, adding/deleting namespaces from an active interleave set will
only be possible via sysfs.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>