nd: calculate a unique cookie per interleave set
An interleave set cookie is the fletcher64 sum of the nfit_spa and
nfit_mem tables that make up an interleave set.  This cookie allows the
kernel to determine if the label on the dimm was created against the
current NFIT configuration.  If the NFIT configuration changes the data
is forfeit and the labels are invalidated.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>