nd: write pmem label set
After 'alt_name', 'uuid', and 'size' have been set to valid values the
labels on the dimm can be updated.

Write procedure is:
1/ Allocate and write new labels in the "next" index
2/ Free the old labels in the working copy
3/ Write the bitmap and the label space on the dimm
4/ Write the index to make the update valid

Once labels are established arrange for the pmem driver to claim the
related dpa ranges on the dimm.  This is a second level of verification
to prevent PMEM vs BLK aliasing.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>