IB/hfi: Have link up/down control SDMA go running.
Advance the idea that the SDMA engines will be in the
running state when the link is up and the idle state
when the links are down.

This mod adds calls to send go_running and go_idle events
when the link goes up or down, respectively.

Presently, a go_idle event does nothing.  This commit does
not change that.

INTERNAL: PR 126102 - Recover from SDMA halt errors

Change-Id: I1bd94145c77435fc20521f5d8749302fcbdae9af
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>