IB/hfi: Add hardware halt wait
Add a wait for the hardware halt to be completed.

When kicking off an SDMA engine, the driver sets the SDMA
halt bit and waits for a halt interrupt before proceeding.
This is incorrect.  The interrupt is to inform the driver
that a halt has occured, not that it is completed.  The
driver must wait for the halt completed bit before
proceeding.

Details:
o Add a work struct and wait routine that will wait for the
  engine to be halted.
o Turn off intterrupt enable in the s10 state.  It is not needed.
o Rename event sdma_event_e15_hw_started1 to
  sdma_event_e15_hw_halt_done to better reflect its purpose.
o Remove the state machine code in sdma_engine_error().  It
  doesn't belong there and the driver is not waiting for the
  interrupt anyway.
o In the state machine, add a call to kick off the worker to
  wait.

INTERNAL: PR 126102 - Recover from SDMA halt errors

Change-Id: I4d1fb5dee59813a3b996bf144ff72f8b40541475
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>