IB/hfi: constrain num_vls by SDMA engines
The send side QOS objective is to never have
a lower priority packet before a higher priority
packet in any of the egress mechanisms.
    
For send contexts, this is fine, but for SDMA engines
the FPGA device limits us to 4.
    
This patch clamps the num_vls module parameter to
4 if the chip reports a smaller value than the
specified number of vls and sdma is enabled.
    
To facilitate this patch, the chip sizing registers
are read before the fixup.
    
INTERNAL: PR 125291 - Host crashes when use_sdma=1 hfi param
INTERNAL: is set and num_vls=8 and SM is started
    
Change-Id: I8986b04891730c2e07c05e10c7abce9ca2970d4e
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>