IB/hfi: Only go to Polling if cable module is present
When the cable is pulled out of the port, WFR transitions
to Offline state and then moves to Polling state instead
of remaining in Offline state. This patch checks for the
presence of the module before re-enabling the link and
if not present, prevents link bringup, and schedules a
link restart.

INTERNAL: PR 124909 - HCA goes to Polling instead of Offline
INTERNAL: when cable is disconnected

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Easwar Hariharan <easwar.hariharan@intel.com>
Change-Id: I59b929ae3b549a154ccf79070314109a8105ec13