IB/hfi: Disables the DLID checking when snooping.
This patch disabled the DLID checking for ingressed packets that normally
happens based on the LID and LMC value. When operating in snoop mode we need
to be able to land all packets. Once the snoop device is closed the
previously used values will be returned to the register.

INTERNAL: PR 126116 - Snoop device should disable ingress DLID-checking.

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: I414bd809b9876c5291b9c0f571b82430e9a69b9a