IB/hfi/test: Unit/regression test for WFR Cntrs
This patch adds a new test script to our test collection. It will collect
counters with iba_pmaquery, hfistats, and hfidiags. Compare them to each other
and ensure they are in agreement. It does this after running an MPI job and
manually forcing a Tx/Rx flit counters to cross the 32bit limit and 64bit
limit. It will then clear the counters and run the tests again. If the counters
did not clear then the test would fail.

INTERNAL: PR 124693 - Counters exported to FM from WFR/Driver need to be 64bit
INTERNAL: and should not wrap when saturated.
INTERNAL: See design note: DN_synthetic_cntrs.docx

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: I5e850e81d38ee0741e5359c1bebf916452310e2d