nd_pmem: Fix sparse warnings in another way
0day has been reporting various errors regarding the use of __iomem with
the "virt_addr" variable of the pmem_device structures.  A few previous
attempts to address this have added __iomem to virt_addr, a bunch of
functions that return pointers within the virt_addr space, to local
variables within this space, etc.  I was still getting one sparse error
about how we were using the return value from pmem_lookup_pg_addr() in
pmem_direct_access(), and it made me realize that we may have been going
about these warnings the wrong way.

My understanding of __iomem is that it essentially annotates the pointer
saying that it shouldn't be dereferenced.  virt_addr points to
persistent memory, though, which can be dereferenced, so this annotation
seems incorrect.

Instead, this patch just casts the interactions between virt_addr (still
a void*) and ioremap_cache() and iounmap(), which expect an __iomem
void*.  That let's us keep virt_addr as directly usable, and gets rid of
our sparse warnings.

These changes revert the following two commits:

873b883cae03 ("nd: fix sparse errors in pmem")
dde04dcde13f ("nd: pmem, just a couple more annotations")

Dan Williams pointed out that the real root of this issue is that it
seems like the memory returned by ioremap_cache() probably shouldn't
have the __iomem annotation.  It looks like most if not all of the uses
of ioremap_cache() either force cast away the __iomem annotation, or
just ignore it and assign to a void*, which will cause sparse warnings.
The real fix most likely is to send out a patch which changes the return
type of ioremap_cache(), and then remove the force casts in our code and
other places in the kernel.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>