IB/hfi: Add SDMA freeze handling code
This adds SDMA code to handle an SPC freeze.  The code is not
yet called.

SDMA freeze handling adds two new states, two new events, and
swaps two states in the SDMA engine state machine.

The new state HWFreeze is a holding state between running and
active clean-up.

The new state FreezeSwClean waits for the software clean to
finish before proceeding.

The new events are a freeze notification and an unfreeze
notification.

When going into an SPC freeze, each engine will receive a freeze
event.  This event will pull the engine out of running into a
hardware freeze holding state.  When in this state, new descriptor
entries rejected.  Each engine will remain in this state until it
is notified that the freeze has been cleared in the hardware.  The
engine will then start a software clean which will clear all in-process
descriptors with an error, then move back to the running state.

The state order swap changes the engine clean-up sequence from
        halt wait
        hw clean
        sw clean
to
        halt wait
        sw clean
        hw clean

When retiring finished descriptors after an engine error or SPC
freeze, take care not to read an invalid hardware head.  For an
engine error, do the software clean before the hardware clean
so a valid hardware head is availalbe.  For an SPC freeze, we
clean after the hardware has been unfrozen, so the harwdare
head has been reset and nothing can be retired.

INTERNAL: PR 124782 - Recover from Freeze mode per requirements
INTERNAL: specified in WFR HAS

Change-Id: I45e91338e0ce695cfc005901906f94abe8fafe41
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>