IB/hfi: Add erratum 291201 workaround
The workaround it to kill a live link if two LCB MBE
errors occur.  To do this, add the bits to a
kill-the-link-on-this-error LCB register.

INTERNAL: PR 125113 - Erratum 291201 - Uncorrectable error for
INTERNAL: an even length 9B packet in combination with AU boundary
INTERNAL: crossing can lead to a potential credit mismatch

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>