IB/hfi: Add LNI Z value
This follows changes to the STL spec where a new LNI fabric
capabilities field, Z, was added.  Add the field, and the
associated changes.

In short: WFR will advertize Z=1, meaning it doesn't support
an AU transmit of 8 bytes.  If the peer advertizes a recive
AU of 8 bytes (the default for FXR, APR) then both sides will
need to move to an AU if 16.

INTERNAL: PR 124549 - Map received vAU of 0 to a 16 byte AU
INTERNAL: in WFR. Also set the Z bit in the advertised Fabric
INTERNAL: Layer Capabilities LNI frame.

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>