DEBUG nfit_test: add 3rd NFIT to test fake AEPs
This patch adds a third NFIT to nfit_test, hard coding the values for
our simulated single AEP DIMM config in Simics.

This patch shouldn't be included in the patch series that we plan on
releasing externally.

I added a mechanism that keeps this third NFIT from being loaded on
non-Simics systems by looking at the x86_model number.  We could do
something more precise like looking for Simics's broken NFIT, but this
would require us to either make nd_acpi.ko a dependency of nfit_test.ko,
or to repeat or make common a bunch of NFIT discovery code.  The
x86_model test is a bit simpler and should serve our purpose for the
limited life of this test code, but I'm open to suggestions if there is
a better way.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>