IB/hfi: free 'cc_state' on driver unload
There's a memory leak on driver unload - we're not freeing the
'cc_state' structure. This is an RCU protected structure, so make
sure that we wait until the RCU callback (which really frees the
memory) has finished before we let the module's code go away.

INTERNAL: PR 125898 - hfi leaks 'cc_state' on driver unload

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: Id91f41222bac151be5ebdf30cb44058c849a14d5