STL core client interface
This patch simplifies the STL client interface so that clients
do not need to pass in a device_id_table. stl_core_client_register(..)
now (re)uses the subsys_interface_register(..) calls in the kernel
to register a client for an STL device. Clients get an add callback
when a device is added and a remove callback when a device is removed.

The character device driver is updated to use the modified interface
along with some minor cleanups in stl_core.

Reviewed-by: Brian Welty <brian.welty@intel.com>
Signed-off-by: Sudeep Dutt <sudeep.dutt@intel.com>