IB/hfi: run opcode counter test as root
This fixes an issue where if the tests are run as root then
in turn run as a normal use the test will fail.

This patch makes the test run as root to insure the
file created is always owned by root to avoid the issue.

INTERNAL: PR 126234 - OpcodeCounters regression test fails if
INTERNAL: user is changed

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>