IB/hfi: Recognize offline state in SnoopIoctl test
A recent driver change exposed the STL physical state
offline.  The link can be offline while the snoop ioctl
test is running, leading to the occasional failure if
the test asked for the state at the wrong time.

Fix this by recognizing the new state and adding it
to cases where it is a valid intermediate state.

INTERNAL: PR 126273 - SnoopIoctl.py sometimes fails

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>