nd_blk: Start using struct ndbw_device
Previously the nd block window driver relied on globals for the physical
and virtual addresses of the block window apertures, control and status
registers.  Move all of this logic into the ndbw_device so that we can
eventually deal with multiple NVDIMMs in a system.

This commit also improves the method by which we allocate and remove
devices, moving to a probe/remove system used by nd.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>