nd_blk: use multiple block windows per NVDIMM
This commit adds support for using multiple block windows per NVDIMM.  A
block window consists of an aperture, a control register and a status
register.  Eventually we will get the number of block windows per DIMM
from ND as part of our initialization, but right now it's just hard
coded.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>