IB/hfi: correct vl processing for PSM/SDMA
The code that fetched the vl was fetching the vl from
the template request from the wrong spot in the template LRH.

This patch makes the following corrections:
as well as the following:
- The VL is now fetched from the correct spot in the KDETH header template
- The VL is now validated when fetched from the KDETH header template
- The sdma engine and data structures have been re-worked
  (See below)
- The code to handle the "num_sdma" module parameter has been
  moved into sdma.c
- The module pararmeter validation is hardened to prevent
  issues with mismatched num_sdma and num_vls
- The qp to engine selection is adjusted for qos
- The seq lock for the vl to sde mapping has been converted to RCU

Commit 364b5875fe22 ("IB/hfi: constrain num_vls by SDMA engines") already
insured the num_vls cannot be configured beyond the number of sdma
engines.

The underlying mapping data structures now support a non-uniform
loading of engines per vl and arbitrary choices for num_vls and num_sdma.

As such, the sdma_map_init() is modified to accept an optional array
of engines by vl.  NULL implies an automatic "uniform" allocation
with the addition of handling non-power of 2 num_sdma.

sdma_select_engine_vl() is modified to use the new data structures as
well as using rcu read locking.

sdma_exit() uses an RCU barrier to gate the free of any mapping.

There is not yet any use of the non-uniform loading.  All callers of
sdma_map_init() are modified to add the NULL as well as dealing with
the potential failure case.

INTERNAL: PR 125510 - mpi_stress/psm hangs with Oops while trying to use
INTERNAL: multiple sdma engines
INTERNAL: PR 125959 - Enhance hfi driver to support multiple SDMA
INTERNAL: engines per VL

Change-Id: I67827aadb392cc339a2597b1e022e23966ed011c
Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>