Ib/hfi: Update sysfs files read by hfi_control to have valid values
This patch fills in missing values in sysfs files which are read by
hfi_control and possibly other IB utils. Some of the formatting may need
tweaking in the future and the WFR string will need to be removed but for now
that is the best we can do namewise.

INTERNAL: PR 125309 - fix up sysfs files read by hfi_control

Change-Id: I81c79c8cbc861cfa73383fa422b31b51cf5ae0be
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>