nd: add infrastructure for per-dimm ioctl support
Latest NFIT exposes an NFIT per-dimm in addition to the bus-wide entry
point.  As a consequence we now need an character device per-dimm.  The
existing device model name of dimmX is too generic for a global node in
/dev, so change dimmX to nvdimmX.  The nvdimmX name may change in the
future to memdevX as it appears NFIT will describe volatile memory
devices in the future.

Tested-by: Ross Zwisler <ross.zwisler@linux.intel.com>