nd: per-dimm ->nfit_ctl()
With a DSM entry per-dimm there is no longer any need to carry the
nfit_handle in the ioctl interface, it is implied by the fact that the
operation is targeted to an nvdimmX device.  How an NFIT-provider does
the lookup from an 'nd_dimm' to its internal message passing interface
is provider-specific.  For this reason, we need a nfit_desc->add_dimm()
callback to initialize those details at dimm discovery time.  For the
ACPI provider ->add_dimm() is an acpi_find_child_device() operation to
find dimms by nfit_handle underneath the ACPI0012 device (formerly
ACPI0011).  In addition to not needing nfit_handle at the ioctl
interface it is also not needed at the DSM interface.  Sammy says he
plans to remove it.

These changes in aggregate make it that much more difficult to maintain
compat hacks in the same file.  A later patch will bring the hacks back
in it's own standalone driver for the ACPI0010 device.

Tested-by: Ross Zwisler <ross.zwisler@linux.intel.com>