Ib/hfi/test: Update driver team software to use opafm instead of ifs_fm
Recent changes to IFS packages have included name changes from ifs to opa for
most of the tools. This patch takes care of those changes for the FM. As of
this check-in ifs_fm is no longer a valid option for wfr driver testing.

INTERNAL: PR 126443 - Update driver team software and hosts to handle newly
INTERNAL: named opa packages.

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>