nd: rename nd.c to core.c
This makes it slightly easier to identify the file that houses the core
entry points for the sub-system.  It also follows the pattern of not
repeating the sub-system name in the source file name.

'nd_core.ko' is now 'nd.ko' which visually distinguishes it from the
"nd_*" true-drivers in an lsmod listing.

Tested-by: Ross Zwisler <ross.zwisler@linux.intel.com>