IB/hfi: Read DMA head from CSR upon receipt of an SDMA idle interrupt
The SDMA idle interrupt is not guaranteed to be ordered with respect
to updates to the the dma_head location in host memory. The head
value read might not be fully up to date. If there are pending
descriptors and the SDMA idle interrupt fired then read from the
CSR SDMA head instead to get the latest value from the hardware.
This patch adds a new idle_mask field to sdma_engine structure
to detect if an interrupt source is SDMA idle.

INTERNAL: PR 125502 - Changes required by WFR HAS change for SdmaIdle

Change-Id: If709187568d497d596b2a16e621b315582ca1b77
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Sudeep Dutt <sudeep.dutt@intel.com>