IB/hfi: set disallowed VLs for kernel contexts
Since there is a kernel context per VL, set the 'SendCtxtCheckVL'
register for kernel contexts to disallow packets on all VLs except
the single VL that the context is intended to support. This may
catch some programming errors.

INTERNAL: PR - 125068 - SendCtxtCheckVL is not populated with the
INTERNAL: VLs that are not active

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: If97c4aafcaaf6b849933b32b9754088106453864