nd: fix btt deletion
Previously the unit test was only exercesing the case where the backing
device to a btt instance disappears.  If we delete the btt instance
first it triggers a use-after-free bug due to a failure to clear
->ndio_claim.

While we're in the area add some messages to make the ndio/btt lifecycle
state clearer while debugging.