Ib/hfi: Use correct OUI as specified in hardware.
This patch relies on the GUID which is determined from the hardware, for
determining the OUI. The old method was to hard code the OUI but this is more
flexible. The fix makes the verbs and mad layers use the same value for OUI
whereas before they were different.

INTERNAL: PR 126453 - Return correct OUI on both verbs and mad layer

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: I5eeb7fef80957911f1d1aedfaca91c8a66c905a4