IB/hfi: Add full support for large eager buffers
The HW supports eager buffers of up to 256KB for
each of the Eager RcvArray entries. Up to now, the
largest eager buffer that the driver supported was
32KB.

This commit adds support for the full set of eager
buffer sizes. It also takes away configuration and
control of Eager buffer sizing and allocation from
the user level process and moves it into the driver.

To accomplish the above two changes, the allocation
of the eager buffers is done much earlier in the
receive context's setup steps. The driver has to
also balance the eager buffers in order to satisfy
the requirements of the HW (eager and expected Rcv
Array entry counts are multiples of 8).

Due to the larger size of eager buffers, allocation
of the memory starts at the highest size (256KB) and
is cut in half every time an allocation fail. The
previously allocated buffers are not free, however.
They are just "split" to the new size so the size
of the individual RcvArray entries is kept the same.

INTERNAL: PR124574 - hfi driver to support multi-packets
INTERNAL: per eager buffer

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I3d14ee84cb4d45a6f35fe3052ddf40599674a305