IB/hfi: Defer CLIENT_REREGISTER event until link transition to Active
To support IsActiveOptimize, a link shouldn't require a Set(PortInfo)
to move the link from Armed to Active.  This requires the driver to
defer dispatching IB event for ClientRegister=1 (when Arming) until the
link transitions to Active.
The IB event is dispatched inside set_link_state() to handle the link
Active transition from both a Set(PortInfo) or from the SMA idle message.

INTERNAL: PR 124163 - client reregister needs to be handled in
INTERNAL: conjunction with active optimize

Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Signed-off-by: Brian Welty <brian.welty@intel.com>
Change-Id: Ibcdc678b8dd17cd258bc6b42598d34015ef82789