IB/hfi: handle CNPs as normal (non-error) packets (B0 and later h/w)
On pre-B0 h/w, the CNP opcode was unrecognized by WFR, and so packets
with the CNP opcode were all delivered as error packets to context 0.
B0 and later h/w will recognize the CNP opcode and deliver them as
normal packets.

INTERNAL: PR - 125634 - Erratum 291394 - 9B packets with a BTH.OpCode
INTERNAL: of 0x80 are decoded as reserved opcodes and not as congestion
INTERNAL: notification packets (CNPs)

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: Id2782b4db543ac12df3d20700b9644dc276dd22f