Ib/hfi: Remove BUG_ON from set_lidlmc.
There is a BUG_ON call in set_lidlmc() that trips when the code comes in there
with snoop or capture enabled. The code should not call in there when snoop is
enabled, but can when capture is enabled if the FM is running. This patch
removes the BUG_ON and puts in an innocuous warning message.

INTERNAL: 126693 -  Packet capture while FM is running can lead to a crash.

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>