IB/hfi: Abort freeze recovery in A0 for RxDmaCsrUncErr
If RXDMA detects an uncorrectable error in a RcvHdrAddr or
RcvHdrTailAddr CSR when delivering a packet to host memory, the error
is correctly reported as RxDmaCsrUncErr and leads to SPC freeze. The
problem is that a write TLP can be sent upstream to a corrupted host
memory address. A multi-bit hardware data corruption on this internal
state can lead to host memory corruption along with the normal error
reporting and SPC freeze mode entry. The workaround for A0 is to not
attempt to recover from freeze mode when the RxDmaCsrUncErr error
case occurs. A host reboot is instead required to recover. This
erratum is slated for a B0 fix. This patch adds an abort argument
to start_freeze_handling(..) so that other errors can abort in a
similar manner if required.

INTERNAL: PR 126634 - Erratum 291567 - Uncorrectable error on
INTERNAL: RcvHdrAddr or RcvHdrTailAddr CSRs can cause a write to a
INTERNAL: corrupted host memory address

Change-Id: Icf0beacd1aaaf33e9260ae91300ea8787853312e
Reviewed-by: Dean Luick <dean.luick@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Sudeep Dutt <sudeep.dutt@intel.com>